requirements.txt
absl-py==2.3.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
arabic-reshaper==3.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
asyncio==3.4.3
attrs==24.2.0
Authlib==1.6.0
babel==2.16.0
beautifulsoup4==4.13.0b2
bleach==6.1.0
blinker==1.9.0
boto3==1.38.5
botocore==1.38.5
Brotli==1.1.0
cachetools==5.5.2
catboost==1.2.8
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cirq==1.6.1
cirq-aqt==1.6.1
cirq-core==1.6.1
cirq-google==1.6.1
cirq-ionq==1.6.1
cirq-pasqal==1.6.1
cirq-web==1.6.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.3.0
cycler==0.12.1
cyclopts==3.22.2
Cython==3.1.3
dataclasses-json==0.6.7
debugpy==1.8.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.4.0
dimod==0.12.20
diskcache==5.6.3
distro==1.9.0
dlt==1.13.0
dnspython==2.7.0
docstring_parser==0.16
docutils==0.21.2
duet==0.2.9
dwave-cloud-client==0.13.6
dwave-gate==0.3.4
dwave-hybrid==0.6.14
dwave-inspector==0.5.4
dwave-ocean-sdk==8.4.0
dwave-optimization==0.6.2
dwave-preprocessing==0.6.9
dwave-samplers==1.6.0
dwave-system==1.32.0
dwave_networkx==0.8.18
dwavebinarycsp==0.3.1
email_validator==2.2.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
executing==2.1.0
fake-useragent==2.1.0
fastapi==0.116.1
fastembed==0.7.1
fasteners==0.19
fastjsonschema==2.20.0
fastmcp==2.10.5
filelock==3.18.0
filetype==1.2.0
findlibs==0.0.5
Flask==3.1.0
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.54.1
fqdn==1.5.1
frozenlist==1.7.0
fsspec==2025.5.1
gast==0.6.0
GetOldTweets3==0.0.11
gitdb==4.0.12
GitPython==3.1.44
giturlparse==0.12.0
google-api-core==2.25.1
google-auth==2.40.3
google-pasta==0.2.0
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.14.0
h2==4.2.0
h5py==3.14.0
hexbytes==1.3.1
homebase==1.0.1
hpack==4.1.0
httpcore==1.0.6
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.34.1
humanfriendly==10.0
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.29.0
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jiter==0.10.0
jmespath==1.0.1
joblib==1.4.2
Js2Py_3.13==0.74.1
json5==0.9.25
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpickle==4.0.5
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kagglehub==0.3.11
kaitaistruct==0.10
keras==3.10.0
kiwisolver==1.4.7
kmedoids==0.5.3.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.8
langsmith==0.4.8
lazy_loader==0.4
liac-arff==2.5.0
libclang==18.1.1
lightgbm==4.6.0
llvmlite==0.44.0
loguru==0.7.3
lxml==5.3.2
m3u8==6.0.0
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mcp==1.11.0
mdurl==0.1.2
minorminer==0.2.19
minsearch==0.0.4
mistune==3.0.2
ml_dtypes==0.5.1
mmh3==5.1.0
mne==1.9.0
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
namex==0.1.0
narwhals==1.37.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.28.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numba==0.61.0
numpy==2.1.3
onnx==1.17.0
onnxruntime==1.22.0
openai==1.97.1
openapi-pydantic==0.5.1
opencv-python==4.12.0.88
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.16.0
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathvalidate==3.3.1
pbr==6.1.1
penaltymodel==1.2.0
pendulum==3.1.0
pillow==11.0.0
platformdirs==4.3.6
plotly==5.24.1
plucky==0.4.3
pluggy==1.6.0
ply==3.11
pooch==1.8.2
portalocker==2.10.1
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.2
proto-plus==1.26.1
protobuf==3.20.3
psutil==6.1.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
py_rust_stemmers==0.1.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydivert==2.1.0
Pygments==2.18.0
pyjsparser==2.7.1
pylatexenc==2.10
pyodbc==5.2.0
pyodc==1.4.1
pyOpenSSL==25.0.0
pyotp==2.9.0
pyparsing==3.2.0
PyPDF2==3.0.1
pyperclip==1.9.0
pyquery==2.0.1
pyreadline3==3.5.4
pyriemann==0.8
PySimpleGUI==5.0.4
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==2.0.7
python-multipart==0.0.20
pytz==2024.2
pyvis==0.3.2
pywin32==310
pywinpty==2.0.14
PyYAML==6.0.2
pyzmq==26.2.0
qdrant-client==1.14.3
qiskit==2.1.1
qiskit-aer==0.17.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
requirements-parser==0.13.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.0.0
rich-argparse==1.7.1
rich-rst==1.3.1
rpds-py==0.20.0
rsa==4.9
rustworkx==0.16.0
s3transfer==0.12.0
safetensors==0.5.3
scikeras==0.13.0
scikit-fuzzy==0.5.0
scikit-learn==1.6.1
scikit-learn-extra==0.3.0
scipy==1.15.1
seaborn==0.13.2
selenium==4.31.0
selenium-wire==5.1.0
semver==3.0.4
Send2Trash==1.8.3
sentence-transformers==5.0.0
setuptools==75.2.0
shap==0.46.0
simplejson==3.20.1
six==1.16.0
slicer==0.0.8
smmap==5.0.2
sniffio==1.3.1
snscrape==0.7.0.20230622
snscrape-cyberconnect==0.0.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.41
sqlglot==27.0.0
sse-starlette==2.4.1
stack-data==0.6.3
starlette==0.47.1
stevedore==5.4.1
streamlit==1.45.0
sympy==1.14.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
termcolor==3.1.0
terminado==0.18.1
tf2onnx==1.16.1
tf_keras==2.19.0
threadpoolctl==3.5.0
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomlkit==0.13.3
torch==2.8.0
torchvision==0.23.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.54.0
trio==0.29.0
trio-websocket==0.12.2
twikit==2.3.3
typedunits==0.0.1.dev20250728235222
types-python-dateutil==2.9.0.20241003
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.3.1
ultralytics==8.3.179
ultralytics-thop==2.0.15
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.35.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.8.0
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
webvtt-py==0.5.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
win-precise-time==1.4.2
win32_setctime==1.2.0
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xgboost==2.1.4
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
